# Conflicts:
#	Engine/Engine/Core.cpp
#	Engine/Engine/Core.h

+++ Engine/Engine/Components/Sprite.cpp
@@ -3,10 +3,10 @@
 Sprite::Sprite(glm::vec3 position, glm::vec2 size, unsigned int color)
 	: Renderable2D(position, size, color), position(m_Position)
 {
-	texr.x = round(m_Position.x);
-	texr.y = round(m_Position.y);
-	texr.w = round(m_Size.x);
-	texr.h = round(m_Size.y);
+	texr.x = static_cast<int>(round(m_Position.x));
+	texr.y = static_cast<int>(round(m_Position.y));
+	texr.w = static_cast<int>(round(m_Size.x));
+	texr.h = static_cast<int>(round(m_Size.y));
 }
 
 Sprite::Sprite(glm::vec3 position, glm::vec2 size, Texture* texture)
